+++ src/components/sectionList/filters/filterSelectors/DynamicFilters.tsx
@@ -22,6 +22,7 @@ import {
     DataElementFilter,
     CompulsoryFilter,
     DataDimensionFilter,
+    ProgramFilter,
 } from '.'
 
 type FilterKeyToComponentMap = Partial<Record<ConfigurableFilterKey, React.FC>>
@@ -45,6 +46,7 @@ const filterKeyToComponentMap: FilterKeyToComponentMap = {
     indicatorGroup: IndicatorGroupFilter,
     indicatorGroupSet: IndicatorGroupSetFilter,
     indicatorType: IndicatorFilter,
+    program: ProgramFilter,
     publicAccess: PublicAccessFilter,
     valueType: ValueTypeSelectionFilter,
 }