+++ src/lib/sectionList/filters/filterConfig.tsx
@@ -34,6 +34,7 @@ export const filterParamsSchema = z
         publicAccess: z.array(
             z.string().refine((val) => parseAccessString(val) !== null)
         ),
+        program: zodArrayIds,
         valueType: z.array(z.nativeEnum(DataElement.valueType)),
     })
     .partial()
@@ -61,6 +62,7 @@ export const filterQueryParamType = {
     indicatorGroup: CustomDelimitedArrayParam,
     indicatorGroupSet: CustomDelimitedArrayParam,
     publicAccess: CustomDelimitedArrayParam,
+    program: CustomDelimitedArrayParam,
     valueType: CustomDelimitedArrayParam,
 } as const satisfies QueryParamsConfigMap