+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.45.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.45.0...v0.45.1) (2025-06-02)
+
+
+### Reverts
+
+* Revert "feat: program indicator groups list and form (#570)" (#571) ([e4fd41b](https://github.com/dhis2/maintenance-app-beta/commit/e4fd41b8da96faf9d529bd477b9f171e364d965a)), closes [#570](https://github.com/dhis2/maintenance-app-beta/issues/570) [#571](https://github.com/dhis2/maintenance-app-beta/issues/571)
+
+# [0.45.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.44.3...v0.45.0) (2025-06-02)
+
+
+### Features
+
+* program indicator groups list and form ([#570](https://github.com/dhis2/maintenance-app-beta/issues/570)) ([9a01f57](https://github.com/dhis2/maintenance-app-beta/commit/9a01f576a263551feb8b9f49c22febfb0e775c7e))
+
 ## [0.44.3](https://github.com/dhis2/maintenance-app-beta/compare/v0.44.2...v0.44.3) (2025-05-21)